Update linting rules to apply existing rules to PATCH verb where relevant
We plan to use PATCH more widely, so we need to elevate it's significance in the style guide and ensure the rules that generally apply to PUT/DELETE also apply to PATCH. 

I've fallen short of prescribing any particular way to operate with PATCH (eg. JsonPatch) as that might be appropriate in some niche use cases. For now, this is just general guidance on behaviour.